web-scraper/requirements.txt
requests
beautifulsoup4
google-cloud-secret-manager
google-cloud-storage